Use UTC to determine livePosts
I experienced an issue with posts not showing when defining a time with the date frontmatter. For example, I expected a post with the date property `2020-05-04T09:15:00` to be visible, but it wasn't. This is due to the timezone differences. I have fixed this by using UTC, but this could be configured.